+++ examples/island_cluster_with_vm_router/main.tf
@@ -143,7 +143,7 @@ module "net" {
 
 module "gke" {
   source  = "terraform-google-modules/kubernetes-engine/google//modules/beta-private-cluster"
-  version = "~> 38.0"
+  version = "~> 41.0"
 
   depends_on = [google_compute_instance.vm]